+++ csharpsrc/Tests/AlgorithmsCases/SortingTestCase.cs
@@ -7,6 +7,10 @@ namespace Tests.AlgorithmsCases
     using System;
     using System.Linq;
     using Algorithms.SortingHelpers;
+    using BenchmarkDotNet.Attributes;
+#if RELEASE
+    using BenchmarkDotNet.Running;
+#endif
     using Microsoft.VisualStudio.TestTools.UnitTesting;
 
     /// <summary>
@@ -21,15 +25,32 @@ namespace Tests.AlgorithmsCases
         private const int TestArrayLength = 10000;
 
         /// <summary>
+        /// Initializes a new instance of the <see cref="SortingTestCase"/> class.
+        /// </summary>
+        public SortingTestCase()
+        {
+            var (testSource, expectedResult) = PrepareTestSourceAndResult();
+            this.TestSource = testSource;
+            this.ExpectedResult = expectedResult;
+        }
+
+        /// <summary>
+        /// Gets or sets the test source.
+        /// </summary>
+        public int[] TestSource { get; set; }
+
+        /// <summary>
+        /// Gets or sets the expected result.
+        /// </summary>
+        public int[] ExpectedResult { get; set; }
+
+        /// <summary>
         /// Verifies the bubble sort algorithm whether it is correct.
         /// </summary>
         [TestMethod]
         public void VerifyResultFromBubbleSortAlgorithm()
         {
-            var (testSource, expectedResult) = PrepareTestSourceAndResult();
-            var actualResult = testSource.BubbleSort((element1, element2) => element1 > element2);
-
-            Assert.IsTrue(actualResult.SequenceEqual(expectedResult));
+            Assert.IsTrue(this.GetResultFromBubbleSortAlgorithm().SequenceEqual(this.ExpectedResult));
         }
 
         /// <summary>
@@ -38,10 +59,7 @@ namespace Tests.AlgorithmsCases
         [TestMethod]
         public void VerifyResultFromMergingSortAlgorithm()
         {
-            var (testSource, expectedResult) = PrepareTestSourceAndResult();
-            var actualResult = testSource.MergingSort((element1, element2) => element1 > element2);
-
-            Assert.IsTrue(actualResult.SequenceEqual(expectedResult));
+            Assert.IsTrue(this.GetResultFromMergingSortAlgorithm().SequenceEqual(this.ExpectedResult));
         }
 
         /// <summary>
@@ -50,10 +68,7 @@ namespace Tests.AlgorithmsCases
         [TestMethod]
         public void VerifyResultFromQuickSortAlgorithm()
         {
-            var (testSource, expectedResult) = PrepareTestSourceAndResult();
-            var actualResult = testSource.QuickSort((element1, element2) => element1 > element2);
-
-            Assert.IsTrue(actualResult.SequenceEqual(expectedResult));
+            Assert.IsTrue(this.GetResultFromQuickSortAlgorithm().SequenceEqual(this.ExpectedResult));
         }
 
         /// <summary>
@@ -62,10 +77,7 @@ namespace Tests.AlgorithmsCases
         [TestMethod]
         public void VerifyResultFromInsertionSortAlgorithm()
         {
-            var (testSource, expectedResult) = PrepareTestSourceAndResult();
-            var actualResult = testSource.InsertSort((element1, element2) => element1 > element2);
-
-            Assert.IsTrue(actualResult.SequenceEqual(expectedResult));
+            Assert.IsTrue(this.GetResultFromInsertionSortAlgorithm().SequenceEqual(this.ExpectedResult));
         }
 
         /// <summary>
@@ -74,11 +86,59 @@ namespace Tests.AlgorithmsCases
         [TestMethod]
         public void VerifyResultFromSelectionSortAlgorithm()
         {
-            var (testSource, expectedResult) = PrepareTestSourceAndResult();
-            var actualResult = testSource.SelectionSort((element1, element2) => element1 > element2);
-
-            Assert.IsTrue(actualResult.SequenceEqual(expectedResult));
+            Assert.IsTrue(this.GetResultFromSelectionSortAlgorithm().SequenceEqual(this.ExpectedResult));
         }
+
+        /// <summary>
+        /// Checks the performance of sorting algorithms.
+        /// </summary>
+        [TestMethod]
+        public void CheckSortingAlgorithmsPerformance()
+        {
+#if RELEASE
+            BenchmarkRunner.Run<SortingTestCase>();
+#endif
+        }
+
+        /// <summary>
+        /// Gets the result from bubble sort algorithm.
+        /// </summary>
+        /// <returns>The sorted result by bubble sort algorithm.</returns>
+        [Benchmark]
+        public int[] GetResultFromBubbleSortAlgorithm() =>
+            this.TestSource.BubbleSort((element1, element2) => element1 > element2).ToArray();
+
+        /// <summary>
+        /// Gets the result from quick sort algorithm.
+        /// </summary>
+        /// <returns>The sorted result by quick sort algorithm.</returns>
+        [Benchmark]
+        public int[] GetResultFromQuickSortAlgorithm() =>
+            this.TestSource.QuickSort((element1, element2) => element1 > element2).ToArray();
+
+        /// <summary>
+        /// Gets the result from merging sort algorithm.
+        /// </summary>
+        /// <returns>The sorted result by bubble sort algorithm.</returns>
+        [Benchmark]
+        public int[] GetResultFromMergingSortAlgorithm() =>
+            this.TestSource.MergingSort((element1, element2) => element1 > element2).ToArray();
+
+        /// <summary>
+        /// Gets the result from insertion sort algorithm.
+        /// </summary>
+        /// <returns>The sorted result by insertion sort algorithm.</returns>
+        [Benchmark]
+        public int[] GetResultFromInsertionSortAlgorithm() =>
+            this.TestSource.InsertSort((element1, element2) => element1 > element2).ToArray();
+
+        /// <summary>
+        /// Gets the result from selection sort algorithm.
+        /// </summary>
+        /// <returns>The sorted result by selection sort algorithm.</returns>
+        [Benchmark]
+        public int[] GetResultFromSelectionSortAlgorithm() =>
+            this.TestSource.SelectionSort((element1, element2) => element1 > element2).ToArray();
 
         /// <summary>
         /// Prepares the test source and expected result.