+++ csharpsrc/Tests/LeetCodeCases/No15.ThreeSum/TestCase.cs
@@ -8,6 +8,9 @@ namespace Tests.LeetCodeCases.No15.ThreeSum
     using System.Collections.Generic;
     using System.Linq;
     using BenchmarkDotNet.Attributes;
+#if RELEASE
+    using BenchmarkDotNet.Running;
+#endif
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Questions.LeetCode.No15.ThreeSum;
 
@@ -69,7 +72,7 @@ namespace Tests.LeetCodeCases.No15.ThreeSum
         public void CheckThreeSumSolutionPerformance()
         {
 #if RELEASE
-            BenchmarkRunner.Run<ThreeSumTest>();
+            BenchmarkRunner.Run<TestCase>();
 #endif
         }