+++ csharpsrc/Tests/LeetCodeCases/No3.LongestSubstringWithoutRepeatingCharacters/TestCase.cs
@@ -6,6 +6,9 @@ namespace Tests.LeetCodeCases.No3.LongestSubstringWithoutRepeatingCharacters
 {
     using System;
     using BenchmarkDotNet.Attributes;
+#if RELEASE
+    using BenchmarkDotNet.Running;
+#endif
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Questions.LeetCode.No3.LongestSubstringWithoutRepeatingCharacters;
 
@@ -67,7 +70,7 @@ namespace Tests.LeetCodeCases.No3.LongestSubstringWithoutRepeatingCharacters
         public void CheckLongestSubstringWithoutRepeatingCharactersSolutionPerformance()
         {
 #if RELEASE
-            BenchmarkRunner.Run<LongestSubstringWithoutRepeatingCharactersTest>();
+            BenchmarkRunner.Run<TestCase>();
 #endif
         }