+++ csharpsrc/Tests/LeetCodeCases/No9.PalindromeNumber/TestCase.cs
@@ -6,6 +6,9 @@ namespace Tests.LeetCodeCases.No9.PalindromeNumber
 {
     using System;
     using BenchmarkDotNet.Attributes;
+#if RELEASE
+    using BenchmarkDotNet.Running;
+#endif
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Questions.LeetCode.No9.PalindromeNumber;
 
@@ -60,7 +63,7 @@ namespace Tests.LeetCodeCases.No9.PalindromeNumber
         public void CheckPalindromeNumberSolutionPerformance()
         {
 #if RELEASE
-            BenchmarkRunner.Run<PalindromeNumberTest>();
+            BenchmarkRunner.Run<TestCase>();
 #endif
         }